feat: continue driver scaffolding
Implement the connect() method to turn the JDBC URI and Properties into
usable parameters to later create the SQL Session.

Build up the class scaffolding with the WherobotsSessionSupplier,
responsible for requesting the creation of the SQL Session, wait for it
to be READY, and return a WherobotsSession wrapping the established
WebSocket connection to the session.